+++ benches/basic.rs
@@ -155,35 +155,15 @@ mod cpupool {
 }
 
 mod thread_pool_callback {
-    use chocolates::thread_pool::callback::{Handle, RunnerFactory, Task};
+    use chocolates::thread_pool::callback::{Handle, SimpleThreadPool, Task};
     use chocolates::thread_pool::{Config, GlobalQueue};
     use num_cpus;
     use std::sync::atomic::AtomicUsize;
     use std::sync::atomic::Ordering::SeqCst;
     use std::sync::{mpsc, Arc};
 
-    use chocolates::thread_pool::SchedUnit;
-    use crossbeam_deque::Steal;
-    struct GQueue(crossbeam_deque::Injector<SchedUnit<Task<GQueue>>>);
-    impl GlobalQueue for GQueue {
-        type RawTask = Task<GQueue>;
-        type Task = Task<GQueue>;
-
-        fn steal_batch_and_pop(
-            &self,
-            local_queue: &crossbeam_deque::Worker<SchedUnit<Self::Task>>,
-        ) -> Steal<SchedUnit<Self::Task>> {
-            crossbeam_deque::Injector::steal_batch_and_pop(&self.0, local_queue)
-        }
-        fn push_raw_task(&self, raw_task: SchedUnit<Self::RawTask>) {
-            self.0.push(raw_task);
-        }
-    }
-
     pub fn spawn_many(b: &mut criterion::Bencher) {
-        let pool = Config::new("test-pool").spawn(RunnerFactory::new(), || {
-            GQueue(crossbeam_deque::Injector::new())
-        });
+        let pool = SimpleThreadPool::from_config(Config::new("test-pool"));
 
         let (tx, rx) = mpsc::sync_channel(10);
         let rem = Arc::new(AtomicUsize::new(0));
@@ -207,9 +187,7 @@ mod thread_pool_callback {
     }
 
     pub fn yield_many(b: &mut criterion::Bencher) {
-        let pool = Config::new("test-pool").spawn(RunnerFactory::new(), || {
-            GQueue(crossbeam_deque::Injector::new())
-        });
+        let pool = SimpleThreadPool::from_config(Config::new("test-pool"));
         let tasks = super::TASKS_PER_CPU * num_cpus::get();
 
         let (tx, rx) = mpsc::sync_channel(tasks);
