+++ benches/depth.rs
@@ -70,33 +70,13 @@ mod cpupool {
 }
 
 mod thread_pool_callback {
-    use chocolates::thread_pool::callback::{Handle, RunnerFactory, Task};
+    use chocolates::thread_pool::callback::{Handle, SimpleThreadPool, Task};
     use chocolates::thread_pool::{Config, GlobalQueue};
     use criterion::Bencher;
     use std::sync::mpsc;
 
-    use chocolates::thread_pool::SchedUnit;
-    use crossbeam_deque::Steal;
-    struct GQueue(crossbeam_deque::Injector<SchedUnit<Task<GQueue>>>);
-    impl GlobalQueue for GQueue {
-        type RawTask = Task<GQueue>;
-        type Task = Task<GQueue>;
-
-        fn steal_batch_and_pop(
-            &self,
-            local_queue: &crossbeam_deque::Worker<SchedUnit<Self::Task>>,
-        ) -> Steal<SchedUnit<Self::Task>> {
-            crossbeam_deque::Injector::steal_batch_and_pop(&self.0, local_queue)
-        }
-        fn push_raw_task(&self, raw_task: SchedUnit<Self::RawTask>) {
-            self.0.push(raw_task);
-        }
-    }
-
     pub fn chained_spawn(b: &mut Bencher) {
-        let pool = Config::new("chain-spawn").spawn(RunnerFactory::new(), || {
-            GQueue(crossbeam_deque::Injector::new())
-        });
+        let pool = SimpleThreadPool::from_config(Config::new("chain-spawn"));
 
         fn spawn<G>(c: &mut Handle<'_, G>, res_tx: mpsc::Sender<()>, n: usize)
         where
